moving set_nested_value_in_dict to Serializer, renaming module
