+++ src/pydase/utils/helpers.py
@@ -272,89 +272,6 @@ def get_nested_value_from_DataService_by_path_and_key(
         return current_data.get(key, None)
 
 
-def set_nested_value_in_dict(data_dict: dict[str, Any], path: str, value: Any) -> None:
-    """
-    Set the value associated with a specific key in a dictionary given a path.
-
-    This function traverses the dictionary according to the path provided and
-    sets the value at that path. The path is a string with dots connecting
-    the levels and brackets indicating list indices.
-
-    Args:
-        cache (dict): The cache dictionary to set the value in.
-        path (str): The path to where the value should be set in the dictionary.
-        value (Any): The value to be set at the specified path in the dictionary.
-
-    Examples:
-        Let's consider the following dictionary:
-
-        cache = {
-            "attr1": {"type": "int", "value": 10},
-            "attr2": {
-                "type": "MyClass",
-                "value": {"attr3": {"type": "float", "value": 20.5}}
-            }
-        }
-
-        The function can be used to set the value of 'attr1' as follows:
-        set_nested_value_in_cache(cache, "attr1", 15)
-
-        It can also be used to set the value of 'attr3', which is nested within 'attr2',
-        as follows:
-        set_nested_value_in_cache(cache, "attr2.attr3", 25.0)
-    """
-
-    parts = path.split(".")
-    current_dict: dict[str, Any] = data_dict
-    index: Optional[int] = None
-
-    for attr_name in parts:
-        # Check if the key contains an index part like '[<index>]'
-        if "[" in attr_name and attr_name.endswith("]"):
-            attr_name, index_part = attr_name.split("[", 1)
-            index_part = index_part.rstrip("]")  # remove the closing bracket
-
-            # Convert the index part to an integer
-            if index_part.isdigit():
-                index = int(index_part)
-            else:
-                logger.error(f"Invalid index format in key: {attr_name}")
-
-        current_dict = cast(dict[str, Any], current_dict.get(attr_name, None))
-
-        if not isinstance(current_dict, dict):
-            # key does not exist in dictionary, e.g. when class does not have this
-            # attribute
-            return
-
-        if index is not None:
-            if 0 <= index < len(current_dict["value"]):
-                try:
-                    current_dict = cast(dict[str, Any], current_dict["value"][index])
-                except Exception as e:
-                    logger.error(f"Could not change {path}. Exception: {e}")
-                    return
-            else:
-                # TODO: appending to a list will probably be done here
-                logger.error(f"Could not change {path}...")
-                return
-
-        # When the attribute is a class instance, the attributes are nested in the
-        # "value" key
-        if (
-            current_dict["type"] not in STANDARD_TYPES
-            and current_dict["type"] != "method"
-        ):
-            current_dict = cast(dict[str, Any], current_dict.get("value", None))  # type: ignore
-
-        index = None
-
-    # setting the new value
-    try:
-        current_dict["value"] = value
-    except Exception as e:
-        logger.error(e)
-
 
 def convert_arguments_to_hinted_types(
     args: dict[str, Any], type_hints: dict[str, Any]
